laptop/requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.2
peewee==3.17.8
pillow==11.1.0
